Load firebase data

+++ lib/view/ui/group_screen.dart
@@ -11,9 +11,8 @@ import '../../view/shared/widgets/toast_helper.dart';
 // ignore: must_be_immutable
 class GroupScreen extends StatelessWidget {
   late int groupIndex;
-  GroupScreen(int index) {
+  GroupScreen(int index, {Key? key}) : super(key: key) {
     groupIndex = index;
-    print(groupIndex);
   }
 
   @override
@@ -26,57 +25,59 @@ class GroupScreen extends StatelessWidget {
         return Scaffold(
             appBar: AppBar(
               actions: [
-                state is GetGroupLinkLoading
-                    ? Center(
-                        child: CircularProgressIndicator(
-                        color: Colors.white,
-                      ))
-                    : IconButton(
-                        onPressed: () {
-                          var id = cubit.groups![groupIndex].id;
-                          Clipboard.setData(ClipboardData(
-                              text: "https://docs.google.com/spreadsheets/d/" +
-                                  id +
-                                  "/edit?usp=sharing"));
-                          showToast("The sheet Link copied to clipboard",
-                              type: ToastType.success);
-                        },
-                        icon: Icon(Icons.link),
-                        iconSize: 30,
-                      ),
-                state is DeleteGroupLoading
-                    ? CircularProgressIndicator(
-                        color: Colors.white,
-                      )
-                    : IconButton(
-                        onPressed: () {
-                          customChoiceDialog(context,
-                              title: "Warning",
-                              content: "Are you sure you want to delete user ",
-                              yesFunction: () {
-                            cubit.deleteGroup(groupIndex, context);
-                          });
-                        },
-                        icon: Icon(Icons.restore_from_trash_outlined),
-                        iconSize: 30,
-                      )
+                if (state is GetGroupLinkLoading)
+                  const Center(
+                      child: CircularProgressIndicator(
+                    color: Colors.white,
+                  ))
+                else
+                  IconButton(
+                    onPressed: () {
+                      var id = cubit.groups![groupIndex].id;
+                      Clipboard.setData(ClipboardData(
+                          text: "https://docs.google.com/spreadsheets/d/" +
+                              id +
+                              "/edit?usp=sharing"));
+                      showToast("The sheet Link copied to clipboard",
+                          type: ToastType.success);
+                    },
+                    icon: const Icon(Icons.link),
+                    iconSize: 30,
+                  ),
+                if (state is DeleteGroupLoading)
+                  const CircularProgressIndicator(
+                    color: Colors.white,
+                  )
+                else
+                  IconButton(
+                    onPressed: () {
+                      customChoiceDialog(context,
+                          title: "Warning",
+                          content: "Are you sure you want to delete user ",
+                          yesFunction: () {
+                        cubit.deleteGroup(groupIndex, context);
+                      });
+                    },
+                    icon: const Icon(Icons.restore_from_trash_outlined),
+                    iconSize: 30,
+                  )
               ],
               foregroundColor: Colors.white,
-              shape: RoundedRectangleBorder(
+              shape: const RoundedRectangleBorder(
                 borderRadius: BorderRadius.vertical(
                   bottom: Radius.circular(15),
                 ),
               ),
               title: Text(
                 '${cubit.groups?[groupIndex - 1].name}',
-                style:
-                    TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+                style: const TextStyle(
+                    color: Colors.white, fontWeight: FontWeight.bold),
               ),
             ),
             body: Container(
-                padding: EdgeInsets.all(15),
+                padding: const EdgeInsets.all(15),
                 child: state is GetGroupNamesLoading
-                    ? Center(child: CircularProgressIndicator())
+                    ? const Center(child: CircularProgressIndicator())
                     : ListView.separated(
                         itemCount:
                             cubit.groups?[groupIndex].columnNames?.length ?? 0,
@@ -85,7 +86,7 @@ class GroupScreen extends StatelessWidget {
                               index, context, cubit, groupIndex, state);
                         },
                         separatorBuilder: (context, index) {
-                          return SizedBox(
+                          return const SizedBox(
                             height: 10,
                           );
                         })));
@@ -96,16 +97,16 @@ class GroupScreen extends StatelessWidget {
   Widget groupItemBuilder(int index, BuildContext context, AdminCubit cubit,
       int groupIndex, state) {
     return state is GetGroupPersonLoading
-        ? Center(
+        ? const Center(
             child: Padding(
-              padding: const EdgeInsets.symmetric(vertical: 10),
+              padding: EdgeInsets.symmetric(vertical: 10),
               child: LinearProgressIndicator(minHeight: 10),
             ),
           )
         : Container(
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(10),
-              color: Color.fromRGBO(115, 115, 115, 1.0),
+              color: const Color.fromRGBO(115, 115, 115, 1.0),
             ),
             child: InkWell(
               onTap: () {
@@ -116,24 +117,24 @@ class GroupScreen extends StatelessWidget {
                 children: [
                   Container(
                     color: Colors.orange,
-                    padding: EdgeInsets.all(10),
+                    padding: const EdgeInsets.all(10),
                     child: Text(
                       index < 9 ? '0${index + 1}' : '${index + 1}',
-                      style: TextStyle(
+                      style: const TextStyle(
                           color: Colors.white,
                           fontSize: 25,
                           fontWeight: FontWeight.bold),
                     ),
                   ),
-                  SizedBox(
+                  const SizedBox(
                     width: 10,
                   ),
                   Expanded(
                     child: Padding(
-                      padding: EdgeInsets.all(10),
+                      padding: const EdgeInsets.all(10),
                       child: Text(
-                        '${cubit.groups![groupIndex].studentNames![index]}',
-                        style: TextStyle(
+                        cubit.groups![groupIndex].studentNames![index],
+                        style: const TextStyle(
                             color: Colors.white,
                             fontSize: 20,
                             fontWeight: FontWeight.bold),
